Remove Collider from collide system and tweek ManagerBase
Remove ICollider and colliders logic from collision system, the problem is, I need the sub class in enginePhysic assembly, but colliders class have to be created in Game assembly for have access to component created by the users. It result of a cycling include of assemblies... So I take the initiative of remove this logic from Engine, only remain the entity collision logic.

Add Destroy func for Managers